Add list of probed vertices for future use
Also, optimized access to precepts to iterate over them only once.
IMPORTANT: Inspectors need to be actively probing for the list to be
populated (it is in fact a map with the vertex name and its value
represented as Integer -- if needed change).

Also modify saboteour's behavior to instead of going to nodes with
enemies, attempt an attack.

Signed-off-by: Aharon Campos <aharoncampos@gmail.com>